bai_tap

+++ case_study/case-study/src/main/java/com/model/customer/Customer.java
@@ -9,7 +9,7 @@ import java.util.Set;
 public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(columnDefinition = "int")
+    @Column(columnDefinition = "long")
     private long id;
     @Column(columnDefinition = "varchar(45)")
     private String name;
